Attempting commit outside OneDrive

+++ myaccount/src/shared/header.jsx
@@ -2,17 +2,36 @@ import { Dropdown } from "flowbite-react"
 import { CiUser } from "react-icons/ci";
 import { FaBell } from "react-icons/fa";
 import { Link } from "react-router-dom"
+import axios from "axios"
+import toast from "react-hot-toast";
 
 const Header = () => {
-  const handleLogout = () => {
-    localStorage.clear()
-    window.location.href="http://localhost:5173"
+  const dataToSend = {
+    sessionId: localStorage.getItem("sessionId"),
+    userId: localStorage.getItem("userId"),
   }
+
+  const handleLogout = async () => {
+    try {
+      const response = await axios.post("https://auth.api.rentnasi.com/v2/logout-api", dataToSend)
+      if (response.data.success) {
+        toast.success(response.data.message)
+        localStorage.clear()
+        // window.location.href = "http://localhost:5173"
+        window.location.href="https://auth.rentnasi.com"
+      } else {
+        console.error("Failed to log out:", response.data.message)
+      }
+    } catch (error) {
+      console.error("Error clearing local storage:", error)
+    }
+  }
+
   return (
     <>
       <header className="shadow sticky top-0 z-30 bg-white px-16">
         <nav className="mx-auto container py-2 flex justify-between">
-          <Link to="/">
+          <Link to="/landlord_management_systemdashboard">
             <img
               className="w-[72px] h-[34px]"
               src="/assets/images/rentnasi.png"
@@ -24,10 +43,10 @@ const Header = () => {
           <ul className="flex space-x-5 rounded-xl bg-gray-100 px-4 py-1 text-xs">
 
             <li className="bg-white py-1 px-3 border border-gray-800 rounded">
-              <a href="/">Home</a>
+              <a href="/dashboard">Home</a>
             </li>
             <li className="p-1 rounded-lg">
-              <Link to="/dashboard">
+              <Link to="/view-apps">
                 Apps
               </Link>
             </li>
